+++ tests/test.php
@@ -34,6 +34,7 @@ include_once("creditHolmesLog.php");
 include_once("creditNote.php");
 include_once("creditNoteLog.php");
 include_once("creditPreview.php");
+include_once("creditSigner.php");
 include_once("dynamicBrcode.php");
 include_once("event.php");
 include_once("individualDocument.php");